requirements.txt
# Core Framework
streamlit>=1.28.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# OpenAI
openai>=1.3.0

# Vector Store & Embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Re-ranking
FlagEmbedding>=1.2.0

# HuggingFace Transformers
transformers>=4.35.0
torch>=2.0.0

# Document Processing
pypdf>=3.17.0
unstructured>=0.10.30
tiktoken>=0.5.1

# Text Splitting
langchain-text-splitters>=0.0.1

# Database
sqlalchemy>=2.0.23
pandas>=2.0.0

# Machine Learning
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0

# Optional dependencies
# If you need GPU support for torch, install manually:
# torch with CUDA: pip install torch --index-url https://download.pytorch.org/whl/cu118
